+++ src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -1,6 +1,6 @@
-package kitchenpos.ui;
+package kitchenpos.menu.ui;
 
-import kitchenpos.application.MenuService;
+import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.dto.MenuResponse;
 import org.springframework.http.ResponseEntity;
