+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.menugroup.application;
