+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,6 +1,6 @@
-package kitchenpos.ui;
+package kitchenpos.order.ui;
 
-import kitchenpos.application.OrderService;
+import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.dto.OrderResponse;
 import org.springframework.http.ResponseEntity;
