+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,6 +1,6 @@
-package kitchenpos.ui;
+package kitchenpos.product.ui;
 
-import kitchenpos.application.ProductService;
+import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
