+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,10 +1,8 @@
 package kitchenpos.order.application;
 
-import kitchenpos.application.OrderService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.*;
-import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
