+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.table.application;
 
-import kitchenpos.application.TableService;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
