+++ src/test/resources/application.properties
@@ -1,0 +1,12 @@
+spring.datasource.username=sa
+spring.h2.console.enabled=true
+
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.use_sql_comments=true
+spring.jpa.show-sql=true
+spring.output.ansi.enabled=always
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
+logging.level.org.springframework.jdbc.core=TRACE
